refactor(notifications): use ScreenHeader + grouping hook (VWA-105)
